feat(ui): move node details to expandable rows
