+++ src/components/nodes-table/index.vue
@@ -8,6 +8,8 @@
     :sort-desc.sync="sorting.desc"
     :sort-by.sync="sorting.by"
     :group-by="groupBy"
+    :expanded.sync="expanded"
+    :single-expand="false"
     :single-select="true"
     :value="[selectedNode]"
     @update:group-by="groupBy = $event"
@@ -17,6 +19,7 @@
     :items-per-page.sync="itemsPerPage"
     item-key="id"
     class="elevation-1"
+    show-expand
     show-select
   >
     <template v-slot:top>
@@ -103,6 +106,14 @@
       {{
         item.lastActive ? new Date(item.lastActive).toLocaleString() : 'Never'
       }}
+    </template>
+    <template v-slot:[`expanded-item`]="{ headers, item }">
+      <node-details
+        :headers="headers"
+        :node-actions="nodeActions"
+        :selected-node="item"
+        :socket="socket"
+      ></node-details>
     </template>
   </v-data-table>
 </template>
